move emulator int test config back
