+++ bigtable-hbase-parent/bigtable-hbase-integration-tests/src/test/java/com/google/cloud/bigtable/hbase/KnownEmulatorGap.java
@@ -1,4 +1,4 @@
-package com.google.cloud.bigtable.hbase.categories;
+package com.google.cloud.bigtable.hbase;
 
 /*
  * Copyright 2017 Google Inc. All Rights Reserved.
